Refactor backend environment

+++ backend/main.go
@@ -2,9 +2,11 @@ package main
 
 import (
   "database/sql"
+  "flag"
   "log"
   "net/http"
   "math/rand"
+  "strconv"
   "time"
   "os"
 
@@ -12,18 +14,22 @@ import (
 )
 
 func main() {
-  port := "8081"
-  if len(os.Args) >= 2 {
-    port = os.Args[1]
-  }
-  log.Printf("Listening :%s", port)
+  port := flag.Int("port", 8081, "The exposed port")
+  devMode := flag.Bool("dev", false, "If the server should be started in development mode")
+  flag.Parse()
+
+  log.Printf("Listening :%d", *port)
+  log.Printf("Dev mode is %+v", *devMode)
   rand.Seed(time.Now().UTC().UnixNano())
 
-  fs := http.FileServer(http.Dir("./web/dist"))
-  http.Handle("/", fs)
-  
-  // Clear database
-  os.Remove("tournament.db")
+  if !*devMode {
+    fs := http.FileServer(http.Dir("./web/dist"))
+    http.Handle("/", fs)
+  } else {
+    // Clear database
+    os.Remove("tournament.db")
+  }
+
   _, err := os.Stat("tournament.db")
 
   if os.IsNotExist(err) {
@@ -46,5 +52,5 @@ func main() {
   http.HandleFunc("/api/choice", makeDbClientHandler(choiceRoute, sqliteDatabase))
   http.HandleFunc("/api/result", makeDbClientHandler(resultRoute, sqliteDatabase))
 
-  log.Fatal(http.ListenAndServe(":" + port, nil))
+  log.Fatal(http.ListenAndServe(":" + strconv.Itoa(*port), nil))
 }
